refactor: Rename internal clock client
